requirements.txt
Flask==2.0.0
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1